requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.0
APScheduler==3.10.4
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
openai==1.50.2
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-telegram-bot==21.6
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
python-dotenv==1.0.1
langchain-core==0.3.21
